[refactor] (tron) move common HD test to helper.ts
