+++ packages/coinlib-tron/test/HdTronPayments.test.ts
@@ -1,19 +1,16 @@
 import fs from 'fs'
 import path from 'path'
-import { omit } from 'lodash'
 import { FeeRateType, BalanceResult, TransactionStatus, NetworkType } from '@bitaccess/coinlib-common'
-import { Transaction } from 'tronweb'
-
 import {
-  HdTronPayments, TronTransactionInfo, HdTronPaymentsConfig,
-  TronSignedTransaction, TronUnsignedTransaction,
+  HdTronPayments,
+  TronTransactionInfo,
+  TronSignedTransaction,
 } from '../src'
-
-import { txInfo_209F8, signedTx_valid, txInfo_a0787, signedTx_invalid } from './fixtures/transactions'
 import { hdAccount } from './fixtures/accounts'
+import { waitForExpiration, runHardcodedPublicKeyTests } from './helpers'
 import { END_TRANSACTION_STATES, delay, expectEqualWhenTruthy, logger } from './utils'
 
-const { XPRV, XPUB, PRIVATE_KEYS, ADDRESSES } = hdAccount
+const { XPRV, XPUB, PRIVATE_KEYS } = hdAccount
 
 const EXTERNAL_ADDRESS = 'TW22XzVixyFZU5DxQJwxqXuKfNKFNMLqJ2'
 
@@ -34,158 +31,6 @@ if (fs.existsSync(secretXprvFilePath)) {
   )
 }
 
-const txInfoOmitEquality = ['data.currentBlock', 'confirmations']
-function assertTxInfo(actual: TronTransactionInfo, expected: TronTransactionInfo): void {
-  expect(omit(actual, txInfoOmitEquality)).toEqual(omit(expected, txInfoOmitEquality))
-}
-
-// Wait for the transaction to expire
-async function waitForExpiration(tx: TronUnsignedTransaction) {
-  await delay(Date.now() - (tx.data as Transaction).raw_data.expiration + 100)
-}
-
-function runHardcodedPublicKeyTests(tp: HdTronPayments, config: HdTronPaymentsConfig) {
-  it('getFullConfig', () => {
-    expect(tp.getFullConfig()).toEqual(config)
-  })
-  it('getPublicConfig', () => {
-    expect(tp.getPublicConfig()).toEqual({
-      network: config.network,
-      hdKey: XPUB,
-    })
-  })
-  it('getAccountIds', () => {
-    expect(tp.getAccountIds()).toEqual([XPUB])
-  })
-  it('getAccountId for index 0', () => {
-    expect(tp.getAccountId(0)).toEqual(XPUB)
-  })
-  it('getAccountId for index 10', () => {
-    expect(tp.getAccountId(10)).toEqual(XPUB)
-  })
-  it('getXpub', async () => {
-    expect(tp.getXpub()).toEqual(XPUB)
-  })
-  it('getPayport for index 1', async () => {
-    expect(await tp.getPayport(1)).toEqual({ address: ADDRESSES[1] })
-  })
-  it('getPayport for high index', async () => {
-    expect(await tp.getPayport(10000)).toEqual({ address: ADDRESSES[10000] })
-  })
-  it('resolvePayport resolves for index 1', async () => {
-    expect(await tp.resolvePayport(1)).toEqual({ address: ADDRESSES[1] })
-  })
-  it('resolvePayport resolves for address', async () => {
-    expect(await tp.resolvePayport(ADDRESSES[1])).toEqual({ address: ADDRESSES[1] })
-  })
-  it('resolvePayport resolves for external address', async () => {
-    expect(await tp.resolvePayport(EXTERNAL_ADDRESS)).toEqual({ address: EXTERNAL_ADDRESS })
-  })
-  it('resolvePayport resolves for payport', async () => {
-    const payport = { address: ADDRESSES[1] }
-    expect(await tp.resolvePayport(payport)).toEqual(payport)
-  })
-  it('resolvePayport resolves for payport with index', async () => {
-    const payport = { index: 1, address: ADDRESSES[1] }
-    expect(await tp.resolvePayport(payport)).toEqual(payport)
-  })
-  it('resolvePayport throws for invalid address', async () => {
-    await expect(tp.resolvePayport('invalid')).rejects.toThrow()
-  })
-  it('resolveFromTo is correct for (index, index)', async () => {
-    expect(await tp.resolveFromTo(0, 2)).toEqual({
-      fromAddress: ADDRESSES[0],
-      fromIndex: 0,
-      fromExtraId: undefined,
-      fromPayport: { address: ADDRESSES[0] },
-      toAddress: ADDRESSES[2],
-      toIndex: 2,
-      toExtraId: undefined,
-      toPayport: { address: ADDRESSES[2] },
-    })
-  })
-  it('resolveFromTo is correct for external address', async () => {
-    expect(await tp.resolveFromTo(0, EXTERNAL_ADDRESS)).toEqual({
-      fromAddress: ADDRESSES[0],
-      fromIndex: 0,
-      fromExtraId: undefined,
-      fromPayport: { address: ADDRESSES[0] },
-      toAddress: EXTERNAL_ADDRESS,
-      toIndex: null,
-      toExtraId: undefined,
-      toPayport: { address: EXTERNAL_ADDRESS },
-    })
-  })
-  it('resolveFromTo is correct for internal address', async () => {
-    expect(await tp.resolveFromTo(0, ADDRESSES[2])).toEqual({
-      fromAddress: ADDRESSES[0],
-      fromIndex: 0,
-      fromExtraId: undefined,
-      fromPayport: { address: ADDRESSES[0] },
-      toAddress: ADDRESSES[2],
-      toIndex: null,
-      toExtraId: undefined,
-      toPayport: { address: ADDRESSES[2] },
-    })
-  })
-  it('resolveFromTo throws for address as from', async () => {
-    await expect(tp.resolveFromTo(EXTERNAL_ADDRESS as any, 0)).rejects.toThrow()
-  })
-
-  it('get transaction by hash with a fee', async () => {
-    const tx = await tp.getTransactionInfo('209f8dbefe6bbb9395f1be76dfb581b7bb53197d27cb28fbfe6c819b914c140c')
-    assertTxInfo(tx, txInfo_209F8)
-  })
-  it('get transaction by hash without a fee', async () => {
-    const tx = await tp.getTransactionInfo('a078736ab768b34dc06ca9048dddfa73383947aed0d93f1eff2adde4b7254f39')
-    assertTxInfo(tx, txInfo_a0787)
-  })
-  it('fail to get an invalid transaction hash', async () => {
-    await expect(tp.getTransactionInfo('123456abcdef')).rejects.toThrow('Transaction not found')
-  })
-
-  it('get a balance using xpub and index', async () => {
-    expect(await tp.getBalance(1)).toEqual({
-      confirmedBalance: '0',
-      unconfirmedBalance: '0',
-      spendableBalance: '0',
-      sweepable: false,
-      requiresActivation: false,
-      minimumBalance: '0.1',
-    })
-  })
-  it('get a balance using an address', async () => {
-    expect(await tp.getBalance({ address: 'TBR4KDPrN9BrnyjienckS2xixcTpJ9aP26' })).toEqual({
-      confirmedBalance: '0',
-      unconfirmedBalance: '0',
-      spendableBalance: '0',
-      sweepable: false,
-      requiresActivation: false,
-      minimumBalance: '0.1',
-    })
-  })
-  it('should not throw for unused address', async () => {
-    expect(await tp.getBalance(12345678)).toEqual({
-      confirmedBalance: '0',
-      unconfirmedBalance: '0',
-      spendableBalance: '0',
-      sweepable: false,
-      requiresActivation: false,
-      minimumBalance: '0.1',
-    })
-  })
-
-  it('broadcast an existing sweep transaction', async () => {
-    const result = await tp.broadcastTransaction(signedTx_valid)
-    expect(result).toEqual({
-      id: signedTx_valid.id,
-      rebroadcast: true,
-    })
-  })
-  it('broadcast should fail on invalid tx', async () => {
-    await expect(tp.broadcastTransaction(signedTx_invalid)).rejects.toThrow('Transaction has expired')
-  })
-}
 
 describe('HdTronPayments', () => {
   let testsComplete = false
@@ -343,7 +188,7 @@ describe('HdTronPayments', () => {
         while (!testsComplete && (!tx || !END_TRANSACTION_STATES.includes(tx.status) || tx.confirmations === 0)) {
           try {
             tx = await tp.getTransactionInfo(txId)
-          } catch (e: any) {
+          } catch (e) {
             if (e?.message?.includes('Transaction not found')) {
               logger.log('tx not found yet', txId, e.message)
             } else {
@@ -405,7 +250,7 @@ describe('HdTronPayments', () => {
           const tx = await pollUntilEnded(signedTx)
           expect(tx.amount).toEqual(signedTx.amount)
           expect(tx.fee).toEqual(signedTx.fee)
-        } catch (e: any) {
+        } catch (e) {
           if ((e?.message || (e as string)).includes('Validate TransferContract error, balance is not sufficient')) {
             logger.log('Ran consecutive tests too soon, previous sweep not complete. Wait a minute and retry')
           }
